Refactor class data to new name/directory
