+++ src/main/java/com/everneth/emi/commands/devop/DevopCommand.java
@@ -1,0 +1,5 @@
+package com.everneth.emi.commands.devop;
+
+public class DevopCommand
+{
+}
